+++ backend/tests/test_helpers.py
@@ -1,0 +1,13 @@
+import pytest
+from backend.weekly.helpers import get_callback_url
+
+
+@pytest.mark.parametrize(
+    "token,result",
+    [
+        ("token", "http://localhost:3000?token=token"),
+        ("ola", "http://localhost:3000?token=ola"),
+    ],
+)
+def test_get_callback_url(config, token, result):
+    assert get_callback_url(config, token) == result
