+++ backend/tests/test_views.py
@@ -1,13 +1,8 @@
 from unittest.mock import Mock
 
+from backend.spotify.exceptions import CodeNotProvided
+from backend.weekly import views
 from flask import url_for
-from spotify.exceptions import CodeNotProvided
-from weekly import views
-
-
-def test_homepage(client):
-    response = client.get("/")
-    assert response.status_code == 200
 
 
 def test_authenticate(client):
@@ -24,6 +19,9 @@ def test_callback_will_try_to_retrieve_code(monkeypatch, client, app):
     url = "/callback?code=12345"
     retrieve_method = Mock(return_value="")
     monkeypatch.setattr(views, "retrieve_code", retrieve_method)
+    authorization_client = Mock()
+    authorization_client.get_access_token = Mock(return_value="abc")
+    monkeypatch.setattr(views, "AuthorizationClient", authorization_client)
     response = client.get(url)
     assert retrieve_method.called
     assert response.status_code == 302
