--- tests/spotify/test_api.py
@@ -1,3 +1,0 @@
-from urllib.parse import urlencode
-
-from spotify.consts import SCOPE
